backend/requirements.txt
fastapi
langchain
uvicorn
openai
tiktoken
psycopg2
python-dotenv